Heart breaker, dream taker
Checking if hearts are broken seems to work, I think. I don't know if
it's just me though, but the ai seems to have an issue where it always
plays clubs for the first few tricks. Ledsuit gets reset though so I
don't know what the issue is. Still lags at the start of a new hand